+++ src/main/java/com/epi/epilog/app/controller/FallDetectionController.java
@@ -3,8 +3,8 @@ package com.epi.epilog.app.controller;
 import com.epi.epilog.app.domain.member.Member;
 import com.epi.epilog.app.dto.CommonResponseDto;
 import com.epi.epilog.app.repository.MemberRepository;
-import com.epi.epilog.app.service.MapService;
-import com.epi.epilog.app.service.SMSService;
+import com.epi.epilog.app.service.fall.MapService;
+import com.epi.epilog.app.service.fall.SMSService;
 import com.epi.epilog.global.exception.ApiException;
 import com.epi.epilog.global.exception.ErrorCode;
 import com.epi.epilog.global.utils.CustomUserDetails;
