+++ src/main/java/com/epi/epilog/app/domain/logs/LogMedication.java
@@ -1,6 +1,5 @@
-package com.epi.epilog.app.domain.log;
+package com.epi.epilog.app.domain.logs;
 
-import io.micrometer.common.lang.Nullable;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -16,17 +15,15 @@ import static lombok.AccessLevel.PROTECTED;
 @Builder
 @NoArgsConstructor(access = PROTECTED)
 @AllArgsConstructor(access = PRIVATE)
-public class LogMood {
+public class LogMedication {
     @Id
-    @Column(name="log_mood_id")
+    @Column(name="log_medication_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="log_id")
     private Log log;
-    @ColumnDefault("false")
-    private boolean detailsState;
-    private String type;
-    @Nullable
-    private String details;
+    private String medicationName;
+    @ColumnDefault("0")
+    private Integer dosage; // 복용량
 }
