+++ src/main/java/com/epi/epilog/app/repository/LogMoodRepository.java
@@ -1,7 +1,7 @@
 package com.epi.epilog.app.repository;
 
-import com.epi.epilog.app.domain.log.Log;
-import com.epi.epilog.app.domain.log.LogMood;
+import com.epi.epilog.app.domain.logs.Log;
+import com.epi.epilog.app.domain.logs.LogMood;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
