+++ src/main/java/com/epi/epilog/app/service/logs/DiabetesQueryService.java
@@ -1,6 +1,6 @@
 package com.epi.epilog.app.service.logs;
 
-import com.epi.epilog.app.domain.log.Log;
+import com.epi.epilog.app.domain.logs.Log;
 import com.epi.epilog.app.domain.member.Member;
 import com.epi.epilog.app.dto.CustomUserInfoDto;
 import com.epi.epilog.app.dto.DiabetesResponseDto;
