+++ src/main/java/com/epi/epilog/app/service/logs/LogCommandService.java
@@ -1,8 +1,8 @@
 package com.epi.epilog.app.service.logs;
 
-import com.epi.epilog.app.domain.log.LogExercise;
-import com.epi.epilog.app.domain.log.LogMood;
-import com.epi.epilog.app.domain.log.Log;
+import com.epi.epilog.app.domain.logs.LogExercise;
+import com.epi.epilog.app.domain.logs.LogMood;
+import com.epi.epilog.app.domain.logs.Log;
 import com.epi.epilog.app.domain.member.Member;
 import com.epi.epilog.app.dto.CommonResponseDto;
 import com.epi.epilog.app.dto.LogsRequestDto;
