+++ src/main/java/com/epi/epilog/global/socket/FallDetectionWebSocketHandler.java
@@ -5,9 +5,9 @@ import com.epi.epilog.app.dto.EmerData;
 import com.epi.epilog.app.dto.SensorData;
 import com.epi.epilog.app.dto.CustomUserInfoDto;
 import com.epi.epilog.app.repository.MemberRepository;
-import com.epi.epilog.app.service.EmergencyService;
-import com.epi.epilog.app.service.FallDetectionService;
-import com.epi.epilog.app.service.SMSService;
+import com.epi.epilog.app.service.fall.EmergencyService;
+import com.epi.epilog.app.service.fall.FallDetectionService;
+import com.epi.epilog.app.service.fall.SMSService;
 import com.epi.epilog.global.exception.ApiException;
 import com.epi.epilog.global.exception.ErrorCode;
 import com.epi.epilog.global.utils.CustomUserDetails;
